Implemented following models and tables in the application -
Cart - This model is used to store the cart Items selected by the students. This is generated by Entity Framework. Whenever student selects any food item and choose to Add it to cart, this model / table will be affected.

FoodCategory - Every food item belongs to some category. Typical category is consisting of Breakfast, lunch, dinner, etc. Food item must belong to some category. This model / Table will be used to store this info.

FoodItem - Food Item is actual food dish which a student is going to order. Each food item will have corrospoding price which will be paid by the user.

MyCanteen DbContext - This is main Database context class of EF. This class will be used to connect with the database and perform CRUD operation on the same.

Order - This model / table will be used when a student places any order. Each order will carry a tax defined for that province.

OrderItem - One order may consist of more than one or two food items, this model/table will be used to store the same.

OrderStatus - order status will have all the status info that  Order will have.

ProgramName - The active students will belong to a program. The students who does have the program will be considered as in active by default.

Student - This model is the main actor in the application. Student will perform all the action throughout the application.